refactor: relocated auth button component in common folder
